Add default values to option details
I suggest to include the default values to the explanations of the single options. When I read them, I always wonder what the default value might be, so I find myself constantly scrolling up and down.

I don’t know if the way I did it here makes most sense:
### Option `default`

There probably is a better way to notate this, so take this PR solely as suggestion.
Thanks for considering.